Refactored client-based tests to live inside the clients themselves. Add RemoteRecordLayerTests harness.
