+++ lionrock-fdb-record-layer-client/build.gradle
@@ -3,6 +3,12 @@ plugins {
     id 'maven-publish'
     id 'jacoco'
     id 'signing'
+    id 'org.springframework.boot' version '2.5.3'
+    id 'io.spring.dependency-management' version '1.0.11.RELEASE'
+}
+
+ext {
+    set('springCloudVersion', "2020.0.3")
 }
 
 repositories {
@@ -17,11 +23,28 @@ dependencies {
     // replace after refactoring.
     implementation 'io.github.panghy.foundationdb:fdb-record-layer-core-pb3:999.3.0-SNAPSHOT'
     implementation 'io.grpc:grpc-stub:1.39.0'
-    implementation 'io.grpc:grpc-netty:1.39.0'
     implementation project(':lionrock-foundationdb-client')
+
+    testImplementation project(path: ':lionrock-foundationdb-server', configuration: 'plain')
+    testImplementation project(path: ':lionrock-foundationdb-server', configuration: 'tests')
+    testImplementation project(':lionrock-fdb-record-layer-client')
+
+    // replace after refactoring.
+    testImplementation 'io.github.panghy.foundationdb:fdb-record-layer-core-pb3:999.3.0-SNAPSHOT'
+    testImplementation 'io.grpc:grpc-stub:1.39.0'
+    testImplementation 'io.grpc:grpc-netty:1.39.0'
+    testImplementation 'org.springframework.boot:spring-boot-starter-test'
+    testRuntimeOnly 'org.springframework.cloud:spring-cloud-starter-sleuth'
+    testRuntimeOnly 'io.github.lognet:grpc-spring-boot-starter:4.5.5'
 
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.7.0'
     testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.7.0'
+}
+
+dependencyManagement {
+    imports {
+        mavenBom "org.springframework.cloud:spring-cloud-dependencies:${springCloudVersion}"
+    }
 }
 
 test {
