+++ lionrock-foundationdb-server/src/test/java/io/github/panghy/lionrock/foundationdb/AbstractGrpcTest.java
@@ -31,7 +31,7 @@ import static org.springframework.boot.test.context.SpringBootTest.WebEnvironmen
  *
  * @author Clement Pang
  */
-@SpringBootTest(webEnvironment = NONE, properties = {"grpc.port=0", "grpc.shutdownGrace=-1",
+@SpringBootTest(webEnvironment = NONE, properties = {"grpc.port=0", "grpc.shutdownGrace=0",
     "logging.level.io.github.panghy.lionrock=DEBUG"})
 public abstract class AbstractGrpcTest {
   @Autowired(required = false)
